Move the OptionContainer wrapper component in too
In to the MenuOptions folder that is. It wraps all of the logic which
handles the interaction so it belongs in there for sure.
